dev: not show bubble menu on add image
